Revamped GIT Organization of Files
In this GitHub project, this assignment involved working on Branching and Merging processes.  My other goal was to organize my notes better.  Providing a better presentation that just submitting 8 pages of git code.  I attached a Word document with the questions that the moderator inquired about as well as "color git codes" to make it a more pleasant experience for the reader.  I also re-designed my README.MD file with bullet points and explained what would be done in the project. Lastly, I also included a compliance statement to make my project more pleasing.